+++ tests/controllers/controller_with_exceptions.py
@@ -53,17 +53,17 @@ class ChildCustomException(CustomException):
 class ControllerWithExceptions:
 
     @winter.route_get('declared_but_not_thrown/')
-    @winter.throws(CustomException)
+    @winter.raises(CustomException)
     def declared_but_not_thrown(self) -> str:
         return 'Hello, sir!'
 
     @winter.route_get('declared_and_thrown/')
-    @winter.throws(CustomException)
+    @winter.raises(CustomException)
     def declared_and_thrown(self) -> str:
         raise CustomException('declared_and_thrown')
 
     @winter.route_get('declared_and_thrown_child/')
-    @winter.throws(CustomException)
+    @winter.raises(CustomException)
     def declared_and_thrown_child(self) -> str:
         raise ChildCustomException('declared_and_thrown_child')
 
@@ -72,16 +72,16 @@ class ControllerWithExceptions:
         raise CustomException('not_declared_but_thrown')
 
     @winter.route_get('declared_but_no_handler/')
-    @winter.throws(ExceptionWithoutHandler)
+    @winter.raises(ExceptionWithoutHandler)
     def declared_but_no_handler(self) -> str:
         raise ExceptionWithoutHandler()
 
-    @winter.throws(CustomException, AnotherExceptionHandler)
+    @winter.raises(CustomException, AnotherExceptionHandler)
     @winter.route_get('exception_with_custom_handler/')
     def with_custom_handler(self) -> str:
         raise CustomException('message')
 
-    @winter.throws(WithUnknownArgumentException)
+    @winter.raises(WithUnknownArgumentException)
     @winter.route_get('with_unknown_argument_exception/')
     def with_unknown_argument_handler(self) -> str:
         raise WithUnknownArgumentException()