+++ winter/web/exceptions/problem_annotation.py
@@ -6,4 +6,3 @@ from .problem_handling_info import ProblemHandlingInfo
 @dataclass(frozen=True)
 class ProblemAnnotation:
     handling_info: ProblemHandlingInfo
-    auto_handle: bool = False