+++ winter/web/exceptions/throttle_exception.py
@@ -3,6 +3,6 @@ from http import HTTPStatus
 from .problem import problem
 
 
-@problem(status=HTTPStatus.TOO_MANY_REQUESTS, detail='Request was throttled', auto_handle=True)
+@problem(status=HTTPStatus.TOO_MANY_REQUESTS, detail='Request was throttled')
 class ThrottleException(Exception):
     pass