+++ winter_openapi/__init__.py
@@ -1,5 +1,9 @@
 from enum import Enum
 
+from winter.web.pagination.limits import MaximumLimitValueExceeded
+from winter.data.pagination import Page
+
+from .annotations.global_exception import register_global_exception
 from .enum_inspector import inspect_enum_class
 from .inspectors import SwaggerAutoSchema
 from .method_arguments_inspector import MethodArgumentsInspector
@@ -12,14 +16,17 @@ from .type_inspection import InspectorNotFound
 from .type_inspection import TypeInfo
 from .type_inspection import inspect_type
 from .type_inspection import register_type_inspector
+from .validators import validate_missing_raises_annotations
 
 
-def setup():
+def setup(allow_missing_raises_annotation: bool = False):
     from drf_yasg.inspectors.field import hinting_type_info
-    from winter.data.pagination import Page
     from .page_inspector import inspect_page
 
+    register_global_exception(MaximumLimitValueExceeded)
     hinting_type_info.insert(0, (Enum, inspect_enum_class))
     register_type_inspector(Page, func=inspect_page)
     register_controller_method_inspector(PathParametersInspector())
     register_controller_method_inspector(QueryParametersInspector())
+    if not allow_missing_raises_annotation:
+        validate_missing_raises_annotations()