+++ winter_openapi/annotations/__init__.py
@@ -1,0 +1,3 @@
+from .global_exception import GlobalExceptionAnnotation
+from .global_exception import global_exception
+from .global_exception import register_global_exception