+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.0.0] - 2020-12-08
+
+### New features
+- winter_openapi add annotation @winter_openapi.global_exception
+- winter_openapi add validation check for missing exceptions 
+### Changed
+- winter.web rename annotation @winter.throws → @winter.raises
+- winter.web remove argument for problem annotation `auto_handle` and define default as True 
+
+
 ## [5.0.0] - 2020-10-25
 
 ### New features